@@ -117,8 +117,8 @@
     "postcss-loader": "2.0.10",
     "prop-types": "15.6.0",
     "query-string": "^5.1.1",
-    "react": "16.2.0",
-    "react-dom": "16.2.1",
+    "react": "^16.14.0",
+    "react-dom": "^16.14.0",
     "react-intl": "2.8.0",
     "react-modal": "3.11.1",
     "react-onclickoutside": "6.7.1",
